Remove version range Support Library version
Using a version range for the Support Library version leads to unpredictable builds and causes apps to be automatically updated to the newest version of the Support Library annotations even if they would otherwise be using an older version. Even for the Annotations alone, there can be [breaking changes](https://issuetracker.google.com/issues/75985001) that this can introduce.

As the Support Library goes through alpha and beta releases tied to new versions of Android, this also can force developers onto alpha or beta builds. For example, right now, the latest published version is `28.0.0-alpha1` - something that most developers do not want to rely on yet.

Gradle will automatically update to the latest version of the Support Library included by any library, so the version range wasn't really adding anything but non-hermetic builds.